backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.0
requests==2.28.1
python-dotenv==0.21.0
pydantic==1.10.7
azure-ai-inference